Add dynamic background image to blog post header
The header element in `Content-BlogPost.liquid` now includes an `id` attribute (`header-background`) to facilitate JavaScript manipulation. A new script has been added to the end of the file that runs when the DOM content is fully loaded. This script checks if the `header-background` element has a background image set. If not, it randomly selects an image from a predefined list and sets it as the background image, enhancing the visual appeal of the blog post header.